+++ src/main/java/com/global/api/entities/enums/CardType.java
@@ -1,37 +1,41 @@
 package com.global.api.entities.enums;
 
-import com.global.api.entities.Card;
-
 public enum CardType {
-	VISA,
-	MC,
-	DISC,
-	AMEX,
-	GIFTCARD,
-	PAYPALECOMMERCE;
+    VISA,
+    MC,
+    DISC,
+    AMEX,
+    GIFTCARD,
+    DEBIT,
+    PAYPALECOMMERCE;
 
 
-	String key;
+    String key;
 
-	CardType(String key) { this.key = key; }
+    CardType(String key) {
+        this.key = key;
+    }
 
-	CardType(){}
+    CardType() {
+    }
 
-	public static CardType getValue(String x) {
-		switch (x) {
-			case "visa":
-				return CardType.VISA;
-			case "mastercard":
-				return CardType.MC;
-			case "discover":
-				return CardType.DISC;
-			case "amex":
-				return CardType.AMEX;
-			case "giftcard":
-				return CardType.GIFTCARD;
-			default:
-				return CardType.PAYPALECOMMERCE;
-		}
-	}
+    public static CardType getValue(String x) {
+        switch (x) {
+            case "visa":
+                return CardType.VISA;
+            case "mastercard":
+                return CardType.MC;
+            case "discover":
+                return CardType.DISC;
+            case "amex":
+                return CardType.AMEX;
+            case "giftcard":
+                return CardType.GIFTCARD;
+            case "debit":
+                return CardType.DEBIT;
+            default:
+                return CardType.PAYPALECOMMERCE;
+        }
+    }
 }